+++ example/counter.ts
@@ -4,4 +4,4 @@ new KingWorld()
 	// Set global mutable state
 	.state('counter', 0)
 	.get('/', ({ store }) => store.counter++)
-	.listen(3000)
+	.listen(8080)