+++ example/custom-response.ts
@@ -17,4 +17,4 @@ new KingWorld()
 	.get('/', () => ({
 		hello: 'world'
 	}))
-	.listen(3000)
+	.listen(8080)