+++ example/guard.ts
@@ -35,4 +35,4 @@ new KingWorld()
 					}
 				})
 	)
-	.listen(3000)
+	.listen(8080)