+++ example/html.ts
@@ -35,4 +35,4 @@ new KingWorld()
 				</body>
 			</html>`
 	)
-	.listen(3000)
+	.listen(8080)